requirements.txt
pydantic-ai
openai
requests
python-dotenv
fuzzywuzzy[speedup]
typing-extensions